+++ src/features/index.ts
@@ -10,3 +10,4 @@
 
 export { default as FileManager } from './file';
 export { default as FileJSON } from './json';
+export { default as FileStream } from './stream';